moved otel logic into tracing mod
